Create instagram_messenger.py
You can use this code to send instagram messages to someone sing python. 
Enter your Username and password in the user_name and pass_word fields respectively before running the file.
Give the username of the person u want to chat and the message in the chat: and message: fields while running the file.